LC-divide-two-integers [Runtime: 1 ms; Memory: 40.8 MB]

+++ math/divide-two-integers.java
@@ -22,10 +22,10 @@ class Solution {
             while (remainder >= absDivisor << (shift + 1)){ // pow returns double type; Need to convert to int
                 shift++;
             }
-            quotient += 1 << shift;
+            quotient += 1L << shift;
             remainder -= absDivisor << shift;
         }
 
-        return sign * quotient;
+        return (int)sign * quotient;
     }
 }
